move util functions used by several services to dedicated file
